Defined a method in word chain class to judge user's answer.
